+++ DMXResource.hpp
